+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Sentence corpus creation
 
+- Add sentence id field (`sent_id`) to generated sentence corpora
 - Processes TEI/XML documents to yield separate chunks for body text and footnotes, with each footnote yielded individually as a separate element
 - Add a method for generating sentence embeddings from sentence corpus